Move became readonly after create
I could have done this the readonly? manner, but as it is not evident when in the callback cycle this method is checked, I put it explicitly in the before_save. If readonly? were to disable setters that might have convinced me to use it, but I'm so tired of getting burned by the implicit 'whens' that are part of ActiveRecord and IOC containers in general - it's much easier to debug code you write in sequence.