Did additional work on FEN, now that its link is visible, and I've clicked it a few times to see that it was sucking !
As a byproduct of this, I probably fixed a subtle castling bug - there neededd to be white and black-specific fields for whether kingside/queenside castlings are still an option! Having had only one field per flank, and not per side as well, probably had the effect of making castling unavailable for you once your opponent castled - oops !

More and more I see the wisdom of each board instance knowing what a FEN would know. FEN is just a perfect model for the state of a board, and there's no reason why each play of a move from a PGN file shouldn't update every FEN field. By definition, any move will alter the FEN of a board, so we have a perfect hash function as well. It's actually quite beautiful - two boards are eql? if their FEN strings are equal - and strategy based on eql? boards will be identical as well, for free :)