Fixed the board updating problem !
Did this mainly by a) refreshing my memory on how i hacked together the client-side implementation of stuff (poorly, of course), and b) envisioning how it *should* work via a short convo with a coworker, Kumar who may be the best sounding board person I've seen at CNU. Basically, the idea should be more oriented in the future to sending back an Object - not just board.to_json, which is a start - but a JS object containing all the state needed to update the board, and then invoking a function to activate that object. Then previous boards can just be a client-side array of all those objects.

Per the JS MVC talk tonight, it wouldn't hurt if all the view parameters (board orientation, etc) were a) persisted client-side (or in gameplay model as a json for backup !) and b) 'applied' to the data objects returned.