Correct(er) implementation of last move highlights.
Note the duplication in the code needed to do this- a more complete client-side board object (via a fuller implementation of Board#to_json) and a relying upon its fields only in the rendering would prevent the need for random one-off global JS variables and the bookkeeping needed to keep them up to date. All_boards would contain an array of objects with the complete data needed to repaint a board at any state (including, possibly the allowed_moves of the board), and as an object it could possibly even contain the code needed to update all client state. (Though in a client-side MVC model, this object would only be the model, and some C would be responsible for applying its data to the V)